Refactor: Move time parsing utility to parserutil
This commit moves the time parsing functionality from `pkg/common` to `pkg/common/parserutil` to better align with the package's responsibility of handling parsing and conversion tasks.

The `ParseTime` function has been renamed to `ParseRFC3339Time` for improved clarity and specificity. A godoc comment has also been added to the function. All related import paths and function calls have been updated accordingly.
